+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v0.32](https://github.com/realityforge/proton/tree/v0.32) (2020-02-02) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.31...v0.32)
+
+Changes in this release:
 
 * Make `AnnotationsUtil.findAnnotationValue()` public.
 * Deprecate `AnnotationsUtil.getAnnotationValue(...)` in favour of `AnnotationsUtil.getAnnotationValueValue(...)` as a future version of the library will add a `AnnotationsUtil.getAnnotationValue(...)` that returns an `AnnotationValue`.